Revert package renaming. Back to old structure
