+++ setup.py
@@ -7,14 +7,22 @@ with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='pypyweconnect',
+    name='volkswagencarnet',
     description='Communicate with Volkswagen WeConnect',
     author='Robin Ostlund',
     author_email='me@robinostlund.name',
     url='https://github.com/robinostlund/volkswagencarnet',
     long_description=long_description,
-    packages=find_packages(),
     long_description_content_type='text/markdown',
+    py_modules=[
+        "vw_connection",
+        "vw_vehicle",
+        "vw_dashboard",
+        "vw_utilities",
+        "const",
+        "__init__"
+    ],
+    provides=["volkswagencarnet"],
     install_requires=list(open("requirements.txt").read().strip().split("\n")),
     use_scm_version=True,
     setup_requires=[
