+++ vw_utilities.py
